requirements.txt
streamlit
fastapi
uvicorn[standard]
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv